pyproject.toml
[project]
name = "honto-capture"
version = "0.1.0"
description = "Capture book pages from the Honto desktop reader on macOS"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pillow>=12.1.1",
    "pyobjc-framework-applicationservices>=12.1; sys_platform == 'darwin'",
    "pyobjc-framework-quartz>=12.1; sys_platform == 'darwin'",
    "pyobjc-framework-vision>=12.1; sys_platform == 'darwin'",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
